Improve bit reader performance by loading four bytes at a time

+++ include/BitReader.h
@@ -44,7 +44,7 @@ class BitReader
 
 		SmackerCommon::FileStream *file;
 
-		uint8_t cache;
+		uint32_t cache;
 		uint32_t nCachedBits;
 
 		void FillCache();
